requirements.txt
pandas>=2.2
numpy>=1.26
python-dateutil>=2.9
rapidfuzz>=3.9
networkx>=3.3
streamlit>=1.36
plotly>=5.22
pyarrow>=16.1
